Created custom useEvents hook and fixed types
